requirements.txt
flask
gunicorn
mysql-connector-python
pandas
openpyxl
werkzeug
